Fix to-do listing's style
The SVGs were growing in size all the time, and
stretching beyond the icon div. Had to do some
adjustments with display flex, and added special
class "eighteen-px-size" to hard-code their size.
Also limited max-size of task-listing(name of the
class) to 70% so as the listing doesn't stretch
too much